@@ -2,9 +2,10 @@
   "name": "svelte-daisy",
   "version": "0.3.0",
   "engines": {
-    "node": ">=18.19.0 <19.0.0",
-    "pnpm": ">=8.12.0 <9.0.0"
+    "node": ">=20.17.0 <21.0.0",
+    "pnpm": ">=9.4.0 <10.0.0"
   },
+  "type": "module",
   "license": "MIT",
   "scripts": {
     "dev": "vite dev",
@@ -12,54 +13,46 @@
     "preview": "vite preview",
     "check": "svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-check --tsconfig ./tsconfig.json --watch",
-    "format": "prettier --cache --write \"**/*.{cjs, js, ts, svelte}\"",
-    "lint": "eslint . --ext .cjs,.js,.ts,.svelte",
-    "lint:fix": "eslint . --fix --ext .cjs,.js,.ts,.svelte",
+    "lint": "pnpm dlx @biomejs/biome check --write ./src",
+    "lint:fix": "pnpm dlx @biomejs/biome check --write ./src",
     "test": "vitest",
     "test:browser": "playwright test",
     "coverage": "vitest run --coverage",
-    "check:all": "pnpm format && pnpm lint:fix"
+    "check:all": "pnpm lint:fix"
   },
   "dependencies": {
-    "@tanstack/svelte-query": "^5.13.4",
-    "@tanstack/svelte-query-devtools": "^5.13.5"
+    "@tanstack/svelte-query": "^5.59.13",
+    "@tanstack/svelte-query-devtools": "^5.59.13"
   },
   "devDependencies": {
-    "@fontsource/fira-mono": "5.0.8",
-    "@fontsource/inter": "^5.0.16",
-    "@fontsource/poppins": "^5.0.8",
+    "@biomejs/biome": "^1.9.4",
+    "@fontsource/fira-mono": "^5.1.0",
+    "@fontsource/mulish": "^5.1.0",
+    "@fontsource/tienne": "^5.1.0",
     "@playwright-testing-library/test": "^4.5.0",
-    "@playwright/test": "^1.40.1",
-    "@sveltejs/adapter-auto": "^3.0.0",
-    "@sveltejs/kit": "^2.0.0",
-    "@sveltejs/vite-plugin-svelte": "^3.0.1",
-    "@tailwindcss/typography": "^0.5.10",
-    "@testing-library/jest-dom": "^6.1.5",
-    "@testing-library/svelte": "^4.0.5",
-    "@types/jest": "^29.5.11",
-    "@typescript-eslint/eslint-plugin": "~6.14.0",
-    "@typescript-eslint/parser": "~6.14.0",
-    "autoprefixer": "^10.4.16",
-    "c8": "^8.0.1",
-    "daisyui": "^4.4.19",
-    "eslint": "~8.55.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-svelte": "^2.35.1",
-    "jsdom": "^23.0.1",
+    "@playwright/test": "^1.48.1",
+    "@sveltejs/adapter-auto": "^3.3.0",
+    "@sveltejs/kit": "^2.7.2",
+    "@sveltejs/vite-plugin-svelte": "~4.0.0",
+    "@tailwindcss/typography": "^0.5.15",
+    "@testing-library/jest-dom": "^6.6.2",
+    "@testing-library/svelte": "^5.2.4",
+    "@types/jest": "^29.5.13",
+    "autoprefixer": "^10.4.20",
+    "bits-ui": "^0.21.16",
+    "c8": "^10.1.2",
+    "daisyui": "^4.12.13",
+    "jsdom": "^25.0.1",
     "playwright-testing-library": "^4.5.0",
-    "postcss": "^8.4.32",
-    "postcss-load-config": "^5.0.2",
-    "prettier": "^3.1.1",
-    "prettier-plugin-svelte": "^3.1.2",
-    "prettier-plugin-tailwindcss": "^0.5.9",
-    "svelte": "5.0.0-next.24",
-    "svelte-check": "^3.6.2",
-    "svelte-preprocess": "^5.1.2",
-    "tailwindcss": "^3.3.6",
-    "tslib": "^2.6.2",
-    "typescript": "~5.3.3",
-    "vite": "^5.0.9",
-    "vitest": "^1.0.4"
-  },
-  "type": "module"
+    "postcss": "^8.4.47",
+    "postcss-load-config": "^6.0.1",
+    "svelte": "~5.0.5",
+    "svelte-check": "^4.0.5",
+    "svelte-preprocess": "^6.0.3",
+    "tailwindcss": "^3.4.14",
+    "tslib": "^2.8.0",
+    "typescript": "~5.6.3",
+    "vite": "~5.4.9",
+    "vitest": "^2.1.3"
+  }
 }
