@@ -1,7 +1,11 @@
 <script lang="ts">
-	import type { Country } from '$lib/models/country';
+import type { Country } from '$lib/models/country'
 
-	export let country: Country;
+interface Props {
+  country: Country
+}
+
+let { country }: Props = $props()
 </script>
 
 <div
@@ -9,13 +13,13 @@
 >
 	<figure><img src={country.flags.svg} alt={country.name.official} /></figure>
 	<div class="card-body">
-		<h2 class="card-title font-poppins text-secondary">{country.name.official}</h2>
-		<p class="font-inter text-xl text-warning">{country.capital ?? ''}</p>
+		<h2 class="card-title font-tienne text-secondary">{country.name.official}</h2>
+		<p class="font-mulish text-xl text-warning">{country.capital ?? ''}</p>
 		{#if country.population > 0}
-			<p class="font-inter text-lg text-info">{country.population.toLocaleString()}</p>
+			<p class="font-mulish text-lg text-info">{country.population.toLocaleString()}</p>
 		{/if}
 		<div class="card-actions justify-end">
-			<button class="btn-neutral btn font-poppins">Details</button>
+			<button class="btn-neutral btn font-tienne">Details</button>
 		</div>
 	</div>
 </div>
