@@ -1,34 +1,37 @@
 <script lang="ts">
-	import { isDark, toggle } from '../stores/theme.store';
+import { isDark, toggle } from '../stores/theme.store'
 
-	const THEMES = {
-		DARK: 'rosepine',
-		LIGHT: 'dawn'
-	};
+const THEMES = {
+  DARK: 'rosepine',
+  LIGHT: 'dawn',
+}
 
-	const toggleTheme = () => {
-		// eslint-disable-next-line svelte/valid-compile
-		document.documentElement.setAttribute('data-theme', $isDark ? THEMES.DARK : THEMES.LIGHT);
-		toggle();
-	};
+const toggleTheme = () => {
+  document.documentElement.setAttribute('data-theme', $isDark ? THEMES.DARK : THEMES.LIGHT)
+  toggle()
+}
 </script>
 
 <div class="form-control">
-	<button class="btn btn-ghost btn-circle" on:click={toggleTheme}>
-		<svg
-			xmlns="http://www.w3.org/2000/svg"
-			fill="none"
-			viewBox="0 0 24 24"
-			stroke="currentColor"
-			width="24px"
-			height="24px"
-		>
-			<path
-				stroke-linecap="round"
-				stroke-linejoin="round"
-				stroke-width="2"
-				d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
-			></path>
-		</svg>
-	</button>
+  <button
+    class="btn btn-ghost btn-circle"
+    onclick={toggleTheme}
+    aria-label="toggle theme"
+  >
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      fill="none"
+      viewBox="0 0 24 24"
+      stroke="currentColor"
+      width="24px"
+      height="24px"
+    >
+      <path
+        stroke-linecap="round"
+        stroke-linejoin="round"
+        stroke-width="2"
+        d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z"
+      ></path>
+    </svg>
+  </button>
 </div>
