@@ -1,23 +1,23 @@
 export interface CountryResponse {
-	countries: Country[];
+  countries: Country[]
 }
 
 export interface Country {
-	cioc: string;
-	capital: string[];
-	name: Name;
-	flags: Flags;
-	population: number;
-	timezones: string[];
-	continent: string[];
+  cioc: string
+  capital: string[]
+  name: Name
+  flags: Flags
+  population: number
+  timezones: string[]
+  continent: string[]
 }
 
 interface Name {
-	common: string;
-	official: string;
+  common: string
+  official: string
 }
 
 interface Flags {
-	png: string;
-	svg: string;
+  png: string
+  svg: string
 }
