@@ -1,10 +1,10 @@
-import { writable } from 'svelte/store';
+import { writable } from 'svelte/store'
 
 // initialize with the light theme, dawn
-const isDark = writable(false);
+const isDark = writable(false)
 
 function toggle() {
-  isDark.update((v) => !v);
+  isDark.update((v) => !v)
 }
 
-export { isDark, toggle };
+export { isDark, toggle }
