@@ -1,8 +1,8 @@
-import { render } from '@testing-library/svelte';
-import Nav from '../lib/components/nav.svelte';
+import { render } from '@testing-library/svelte'
+import Nav from '../lib/components/nav.svelte'
 
 it('has the heading text for "Countries"', () => {
-	const { getByText } = render(Nav);
+  const { getByText } = render(Nav)
 
-	expect(getByText('daisyUI')).toBeTruthy();
-});
+  expect(getByText('daisyUI')).toBeTruthy()
+})
