@@ -1,8 +1,10 @@
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function sort_by_key(array: any[], key: string) {
-	return array.sort(function (a, b) {
-		const x = a[key];
-		const y = b[key];
-		return x < y ? -1 : x > y ? 1 : 0;
-	});
+import type { SafeAny } from './types'
+
+export function sort_by_key(array: SafeAny[], key: string) {
+  return array.sort((a, b) => {
+    const x = a[key]
+    const y = b[key]
+
+    return x < y ? -1 : x > y ? 1 : 0
+  })
 }
