@@ -2,13 +2,13 @@ import { svelte } from '@sveltejs/vite-plugin-svelte';
 import { defineConfig } from 'vite';
 
 export default defineConfig({
-	plugins: [svelte({ hot: !process.env.VITEST })],
-	test: {
-		globals: true,
-		environment: 'jsdom',
-		coverage: {
-			reporter: ['text', 'json', 'html']
-		},
-		setupFiles: ['./src/setupTests.ts']
-	}
+  plugins: [svelte({ hot: !process.env.VITEST })],
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    coverage: {
+      reporter: ['text', 'json', 'html']
+    },
+    setupFiles: ['./src/setupTests.ts']
+  }
 });
